refactor: reorganize PFCF client structure and improve import handling
- Moved PFCFApi class implementation from __init__.py to a new api.py file for better modularity and clarity.
- Updated import statements in app.py to reflect the new structure, enhancing readability.
- Cleaned up unnecessary whitespace and ensured consistent formatting throughout the code.
- These changes streamline the PFCF client architecture, making it easier to maintain and extend.
